+++ src/Traffic.java
@@ -5,7 +5,7 @@ public class Traffic extends Sprite {
     private final int initTraffic_y = -100;
     private int speedOption = 1;
         
-    public Traffic(int x, int y, int speedOption) {
+    public Traffic(double x, double y, int speedOption) {
         super(x, y);
         this.speedOption = speedOption;
         initTraffic();
